reformat: use Abstract Controller to config ExceptionHandler
