+++ shopping-cart-backend/src/main/java/fptu/swp391/shoppingcart/TestController.java
@@ -4,7 +4,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class TestController {
+public class TestController extends AbstractApplicationController{
     @GetMapping("/")
     String test() {
         return "Hello world!";
